refactor: :sparkles: rearanged code and made it possible to run multiple instances

+++ bot/Modules/EconomyModule.cs
@@ -1,13 +1,27 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using StackExchange.Redis;
 using v10.Data.MongoDB;
 
 namespace bot.Modules;
 
 public class EconomyModule : CustomModule<SocketCommandContext>
 {
-    public BotDataService BotDataService { get; set; }
+    public IBotDataService BotDataService { get; set; }
+    public ILogger<EconomyModule> Logger { get; set; }
+
+    public EconomyModule(
+        IServiceProvider serviceProvider,
+        ILogger<EconomyModule> logger
+        )
+    {
+        var server = serviceProvider.GetRequiredService<IServer>();
+        _database = server.Multiplexer.GetDatabase();
+    }
 
     [Command("balance")]
     [Alias("coins")]
@@ -15,18 +29,21 @@ public class EconomyModule : CustomModule<SocketCommandContext>
     [Remarks("balance")]
     public async Task GetBalance()
     {
+        if (!EnsureSingle()) { return; }
         var user = Context.User;
         var guildId = Context.Guild?.Id ?? 0;
         var userId = user.Id;
         var userBalanceResult = await BotDataService.GetUserBalance(guildId, userId);
-        await userBalanceResult.Match(async r => {
+        await userBalanceResult.Match(async r =>
+        {
             var embed = new EmbedBuilder()
                 .WithColor(Color.Blue)
                 .AddField("UserName", $"{user.Username}#{user.Discriminator}")
                 .AddField("Balance", $"{r} :coin:")
                 .Build();
             await ReplyAsync(embed: embed);
-        }, async e => {
+        }, async e =>
+        {
             var embed = new EmbedBuilder()
                 .WithColor(Color.Red)
                 .AddField("UserName", $"{user.Username}#{user.Discriminator}")
@@ -34,6 +51,7 @@ public class EconomyModule : CustomModule<SocketCommandContext>
                 .Build();
             await ReplyAsync(embed: embed);
         });
+        ReleaseLock();
     }
 
     [Command("leave")]
@@ -41,19 +59,22 @@ public class EconomyModule : CustomModule<SocketCommandContext>
     [Remarks("leave")]
     public async Task Leave(IUser user = null)
     {
+        if (!EnsureSingle()) { return; }
         // LeaveEconomy
         user ??= Context.User;
         var guildId = Context.Guild?.Id ?? 0;
         var userId = user.Id;
         var result = await BotDataService.LeaveEconomy(guildId, userId);
-        await result.Match(async r => {
+        await result.Match(async r =>
+        {
             var embed = new EmbedBuilder()
                 .WithColor(Color.Green)
                 .WithTitle("You have left the economy!")
                 .AddField("User", $"{user.Username}#{user.Discriminator}")
                 .Build();
             await ReplyAsync(embed: embed);
-        }, async e => {
+        }, async e =>
+        {
             var embed = new EmbedBuilder()
                 .WithColor(Color.Red)
                 .WithTitle("Failed to leave the economy!")
@@ -62,6 +83,7 @@ public class EconomyModule : CustomModule<SocketCommandContext>
                 .Build();
             await ReplyAsync(embed: embed);
         });
+        ReleaseLock();
     }
 
     [Command("join")]
@@ -69,11 +91,13 @@ public class EconomyModule : CustomModule<SocketCommandContext>
     [Remarks("join")]
     public async Task Join(IUser user = null)
     {
+        if (!EnsureSingle()) { return; }
         user ??= Context.User;
         var guildId = Context.Guild?.Id ?? 0;
         var userId = user.Id;
         var result = await BotDataService.JoinEconomy(guildId, userId);
-        await result.Match(async r => {
+        await result.Match(async r =>
+        {
             var embed = new EmbedBuilder()
                 .WithColor(Color.Green)
                 .WithTitle("Welcome to the economy!")
@@ -81,7 +105,8 @@ public class EconomyModule : CustomModule<SocketCommandContext>
                 .AddField("Balance", $"{r} :coin:")
                 .Build();
             await ReplyAsync(embed: embed);
-        }, async e => {
+        }, async e =>
+        {
             var embed = new EmbedBuilder()
                 .WithColor(Color.Red)
                 .WithTitle("Could not join the economy!")
@@ -90,11 +115,13 @@ public class EconomyModule : CustomModule<SocketCommandContext>
                 .Build();
             await ReplyAsync(embed: embed);
         });
+        ReleaseLock();
     }
 
     [Command("coins")]
     public async Task GetCoins(IUser user = null)
     {
+        if (!EnsureSingle()) { return; }
         user ??= Context.User;
         var guildId = Context.Guild?.Id ?? 0;
         var userId = user.Id;
@@ -105,25 +132,30 @@ public class EconomyModule : CustomModule<SocketCommandContext>
             .AddField("Balance", $"{userData.money} :coin:")
             .Build();
         await ReplyAsync(embed: embed);
+        ReleaseLock();
     }
 
     [Command("addcoins")]
     [RequireUserPermission(GuildPermission.ManageRoles)]
     public async Task AddMoney(IUser user, ulong amount)
     {
+        if (!EnsureSingle()) { return; }
         var guildId = Context.Guild?.Id ?? 0;
         var userId = user.Id;
         var d = BotDataService.AddMoney(guildId, userId, amount);
         await ReplyAsync($"Added **{amount}** :coin: to **{user.Username}**. Their new balance is **{d.money}** :coin:");
+        ReleaseLock();
     }
 
     [Command("removecoins")]
     [RequireUserPermission(GuildPermission.ManageRoles)]
     public async Task RemoveMoney(IUser user, ulong amount)
     {
+        if (!EnsureSingle()) { return; }
         var guildId = Context.Guild?.Id ?? 0;
         var userId = user.Id;
         var d = BotDataService.RemoveMoney(guildId, userId, amount);
         await ReplyAsync($"Removed **{amount}** :coin: from **{user.Username}**. Their new balance is: **{d.money}** :coin:");
+        ReleaseLock();
     }
 }
